+++ derive_builder/src/options/struct_mode.rs
@@ -216,21 +216,7 @@ impl OptionsBuilderMode for StructMode {
 }
 
 impl From<OptionsBuilder<StructMode>> for (StructOptions, OptionsBuilder<FieldMode>) {
-    fn from(mut b: OptionsBuilder<StructMode>) -> (StructOptions, OptionsBuilder<FieldMode>) {
-        // Check if field visibility has been expressly set at the struct level.
-        // If not, and if the crate is operating under the old public fields mode,
-        // present a compilation warning.
-        if !cfg!(feature = "private_fields") && b.field_vis.is_none() {
-            let where_diagnostics = b.where_diagnostics();
-            b.mode.push_deprecation_note(format!(
-                "Builder fields will be private by default starting in the next version. \
-                (see https://github.com/colin-kiegel/rust-derive-builder/issues/86 for \
-                more details). To squelch this message and adopt the new behavior now, \
-                compile `derive_builder` with `--features \"private_fields\"` or add \
-                `field(<vis>)` to the builder attribute on the struct. (Found {})",
-                where_diagnostics));
-        }
-
+    fn from(b: OptionsBuilder<StructMode>) -> (StructOptions, OptionsBuilder<FieldMode>) {
         let field_defaults = OptionsBuilder::<FieldMode> {
             setter_enabled: b.setter_enabled,
             builder_pattern: b.builder_pattern,